+++ blocks/section.ts
@@ -63,7 +63,7 @@ export interface SectionModule<TConfig = any, TProps = any> extends
 
 const sectionBlock: Block<SectionModule> = {
   type: "sections",
-  introspect: {funcNames:["loader", "default"]},
+  introspect: { funcNames: ["loader", "default"] },
   adapt: <TConfig = any, TProps = any>(
     mod: SectionModule<TConfig, TProps>,
     resolver: string,
