+++ engine/schema/gen.ts
@@ -7,13 +7,13 @@ import {
   newSchemaBuilder,
   Schemas,
 } from "$live/engine/schema/builder.ts";
-import { parsePath } from "$live/engine/schema/swc/swc.ts";
-import { programToBlockRef } from "$live/engine/schema/swc/transform.ts";
+import { Schemeable } from "$live/engine/schema/transform.ts";
 import { context } from "$live/live.ts";
 import { TsType } from "https://esm.sh/v130/@swc/wasm@1.3.76/wasm.js";
 import { AppManifest } from "../../blocks/app.ts";
 import { JSONSchema7 } from "../../deps.ts";
-import { Schemeable } from "$live/engine/schema/transform.ts";
+import { parsePath } from "./parser.ts";
+import { programToBlockRef } from "./transform.ts";
 
 export const namespaceOf = (blkType: string, blkKey: string): string => {
   return blkKey.substring(0, blkKey.indexOf(blkType) - 1);
@@ -45,7 +45,7 @@ export const genSchemasFromManifest = async (
   const modulesPromises: Promise<
     (BlockModule | EntrypointModule | undefined)
   >[] = [];
-  const schemeableReferences = new Map<TsType, Schemeable>();
+  const references = new Map<TsType, Schemeable>();
   for (const block of blocks) {
     for (
       const blockModuleKey of Object.keys(
@@ -68,6 +68,7 @@ export const genSchemasFromManifest = async (
               import.meta.resolve(blockModuleKey),
               blockModuleKey,
             ]);
+
       const programPromise = parsePath(blockPath);
       modulesPromises.push(programPromise.then(async (doc) => {
         if (!doc) {
@@ -76,7 +77,7 @@ export const genSchemasFromManifest = async (
         const ref = await programToBlockRef(
           blockKey,
           doc,
-          schemeableReferences,
+          references,
           block.introspect,
         );
         if (ref) {
