+++ engine/schema/schemeable.test.types.ts
@@ -48,7 +48,7 @@ export interface WellKnown {
   // @ts-ignore: "will work as soon as the new engine is entirely here"
   section: InstanceOf<string, "#/root/sections">;
   promiseValue: Promise<string>;
-  resolvable: Resolvable<string>;
+  resolvable: Resolvable;
   // @ts-ignore: "will work as soon as the new engine is entirely here"
   preactComponent: PreactComponent<InstanceOf<string, "#/root/sections">>;
 }
