+++ engine/schema/transform.ts
@@ -1,24 +1,39 @@
 import { JSONSchema7, JSONSchema7Type } from "$live/deps.ts";
-import { beautify, denoDoc, jsDocToSchema } from "$live/engine/schema/utils.ts";
-import { fromFileUrlOrNoop } from "$live/utils/filesystem.ts";
-import {
-  DocNode,
-  InterfaceDef,
-  TsTypeDef,
-  TsTypeDefLiteral,
-  TsTypeLiteralDef,
-  TsTypeParamDef,
-  TsTypeRefDef,
-  TsTypeUnionDef,
-} from "https://deno.land/x/deno_doc@0.58.0/lib/types.d.ts";
-import { JSONSchema7TypeName } from "https://esm.sh/@types/json-schema@7.0.11?pin=102";
-import { crypto, toHashString } from "std/crypto/mod.ts";
-
-export interface TransformContext {
-  base: string;
-  namespace: string;
-}
-
+import { BlockModuleRef, IntrospectParams } from "$live/engine/block.ts";
+import { beautify } from "$live/engine/schema/utils.ts";
+import type { ParsedSource } from "https://denopkg.com/deco-cx/deno_ast_wasm@0.1.0/mod.ts";
+import type {
+  ArrowFunctionExpression,
+  ExportNamedDeclaration,
+  FunctionDeclaration,
+  FunctionExpression,
+  NamedExportSpecifier,
+  NamedImportSpecifier,
+  Param,
+  Pattern,
+  StringLiteral,
+  TsArrayType,
+  TsIndexedAccessType,
+  TsInterfaceDeclaration,
+  TsIntersectionType,
+  TsKeywordType,
+  TsLiteral,
+  TsLiteralType,
+  TsOptionalType,
+  TsParenthesizedType,
+  TsTupleType,
+  TsType,
+  TsTypeAnnotation,
+  TsTypeElement,
+  TsTypeLiteral,
+  TsTypeParameter,
+  TsTypeReference,
+  TsUnionType,
+} from "https://esm.sh/v130/@swc/wasm@1.3.76";
+import { JSONSchema7TypeName } from "https://esm.sh/v130/@types/json-schema@7.0.11/index.d.ts";
+import { dirname, fromFileUrl, join, toFileUrl } from "std/path/mod.ts";
+import { spannableToJsDoc } from "./comments.ts";
+import { parsePath } from "./parser.ts";
 export interface SchemeableBase {
   jsDocSchema?: JSONSchema7;
   friendlyId?: string;
@@ -83,32 +98,407 @@ export type Schemeable =
   | UnknownSchemable
   | SchemeableRef;
 
-const schemeableWellKnownType = async (
-  ref: TsTypeRefDef,
-  root: [string, DocNode[]],
-  seen: Map<DocNode, Schemeable>,
-  typeParams?: TsTypeParamDef[],
-  typeParamsImpl?: Schemeable[],
+export interface SchemeableTransformContext {
+  path: string;
+  parsedSource: ParsedSource;
+  references?: Map<TsType, Schemeable>;
+  instantiatedTypeParams?: Schemeable[];
+  tryGetFromInstantiatedParameters?: (
+    name: string,
+  ) => Promise<Schemeable | undefined>;
+}
+
+const UNKNOWN: UnknownSchemable = {
+  name: "unknown",
+  type: "unknown",
+};
+
+const tsTypeElementsToObjectSchemeable = async (
+  tsTypeElements: TsTypeElement[],
+  ctx: SchemeableTransformContext,
+): Promise<Omit<ObjectSchemeable, "name">> => {
+  const keysPromise: Promise<[string, ObjectSchemeable["value"][string]]>[] =
+    [];
+  for (const prop of tsTypeElements) {
+    if (prop.type !== "TsPropertySignature") {
+      continue;
+    }
+    if (!prop.typeAnnotation) {
+      continue;
+    }
+    const key = prop.key;
+    if (key.type !== "Identifier") {
+      continue;
+    }
+    keysPromise.push(
+      tsTypeToSchemeable(prop.typeAnnotation.typeAnnotation, ctx, prop.optional)
+        .then((
+          schemeable,
+        ) => [
+          key.value,
+          {
+            title: beautify(key.value),
+            jsDocSchema: spannableToJsDoc(prop),
+            schemeable,
+            required: !prop.optional,
+          } as ObjectSchemeable["value"][string],
+        ]),
+    );
+  }
+  const keys = await Promise.all(keysPromise);
+  const schemeable: Omit<ObjectSchemeable, "name"> = {
+    type: "object",
+    value: {},
+  };
+  for (const [key, value] of keys) {
+    schemeable.value[key] = value;
+  }
+  return schemeable;
+};
+
+const getFromParametersFunc = (
+  params: TsTypeParameter[],
+  ctx: SchemeableTransformContext,
+): SchemeableTransformContext["tryGetFromInstantiatedParameters"] => {
+  const {
+    tryGetFromInstantiatedParameters: _tryGetFromInstantiatedParameters,
+    instantiatedTypeParams,
+  } = ctx;
+  const typeParamNameToIdx: Record<string, { idx: number; default?: TsType }> =
+    {};
+
+  for (let paramIdx = 0; paramIdx < params.length; paramIdx++) {
+    const param = params[paramIdx];
+    typeParamNameToIdx[param.name.value] = {
+      idx: paramIdx,
+      default: param.default,
+    };
+  }
+  return async (
+    name: string,
+  ): Promise<Schemeable | undefined> => {
+    const val = typeParamNameToIdx[name];
+    if (!val) {
+      return undefined;
+    }
+    return instantiatedTypeParams?.[val.idx] ??
+      await _tryGetFromInstantiatedParameters?.(name) ??
+      (val.default
+        ? tsTypeToSchemeable(val.default, {
+          ...ctx,
+          tryGetFromInstantiatedParameters: _tryGetFromInstantiatedParameters,
+        })
+        : undefined);
+  };
+};
+const tsInterfaceDeclarationToSchemeable = async (
+  dec: TsInterfaceDeclaration,
+  ctx: SchemeableTransformContext,
+): Promise<Schemeable> => {
+  const {
+    path,
+    tryGetFromInstantiatedParameters: _tryGetFromInstantiatedParameters,
+  } = ctx;
+  const allOfs: Promise<Schemeable>[] = [];
+  const params = dec.typeParams?.parameters ?? [];
+
+  const tryGetFromInstantiatedParameters = getFromParametersFunc(params, ctx);
+  for (const ext of dec.extends) {
+    if (ext.expression.type === "Identifier") {
+      allOfs.push(
+        typeNameToSchemeable(ext.expression.value, {
+          ...ctx,
+          tryGetFromInstantiatedParameters,
+        }),
+      );
+    }
+  }
+  const objectSchemeable = await tsTypeElementsToObjectSchemeable(
+    dec.body.body,
+    {
+      ...ctx,
+      tryGetFromInstantiatedParameters,
+    },
+  );
+  return {
+    ...objectSchemeable,
+    file: path,
+    name: dec.id.value,
+    extends: await Promise.all(allOfs),
+  };
+};
+// cannot have typeParams but can have type parameters on context
+export const typeNameToSchemeable = async (
+  typeName: string,
+  ctx: SchemeableTransformContext,
+): Promise<Schemeable> => {
+  const {
+    path,
+    parsedSource,
+    tryGetFromInstantiatedParameters,
+  } = ctx;
+  const { program } = parsedSource;
+  const val = await tryGetFromInstantiatedParameters?.(typeName);
+  if (val) {
+    return val;
+  }
+  for (const item of program.body) {
+    if (item.type === "ExportNamedDeclaration") {
+      const spec = item.specifiers.find((
+        spec,
+      ) =>
+        spec.type === "ExportSpecifier" &&
+        (spec.exported?.value ?? spec.orig.value) === typeName
+      );
+      if (spec) {
+        const _spec = spec as NamedExportSpecifier;
+        const source = (item as ExportNamedDeclaration).source!.value;
+        const fileUrl = source.startsWith(".")
+          ? join(
+            Deno.cwd(),
+            dirname(path),
+            (item as ExportNamedDeclaration).source!.value,
+          )
+          : source;
+        const from = import.meta.resolve(
+          fileUrl,
+        );
+        const newProgram = await parsePath(from);
+        if (!newProgram) {
+          return UNKNOWN;
+        }
+
+        return typeNameToSchemeable(
+          _spec.exported?.value ?? _spec.orig.value,
+          {
+            ...ctx,
+            path: from,
+            parsedSource: newProgram,
+          },
+        );
+      }
+    }
+    if (
+      item.type === "ExportDeclaration" &&
+      item.declaration.type === "TsInterfaceDeclaration" &&
+      item.declaration.id.value === typeName
+    ) {
+      return {
+        jsDocSchema: spannableToJsDoc(item),
+        ...await tsInterfaceDeclarationToSchemeable(item.declaration, ctx),
+      };
+    }
+    if (
+      item.type === "ExportDeclaration" &&
+      item.declaration.type === "TsTypeAliasDeclaration" &&
+      item.declaration.id.value === typeName
+    ) {
+      const _tryGetFromInstantiatedParameters = getFromParametersFunc(
+        item.declaration.typeParams?.parameters ?? [],
+        ctx,
+      );
+      return {
+        file: path,
+        jsDocSchema: spannableToJsDoc(item),
+        ...await tsTypeToSchemeable(item.declaration.typeAnnotation, {
+          ...ctx,
+          tryGetFromInstantiatedParameters: _tryGetFromInstantiatedParameters,
+        }),
+        name: item.declaration.id.value,
+      };
+    }
+    if (
+      item.type === "TsInterfaceDeclaration" &&
+      item.id.value === typeName
+    ) {
+      return {
+        jsDocSchema: spannableToJsDoc(item),
+        ...await tsInterfaceDeclarationToSchemeable(item, ctx),
+      };
+    }
+    if (
+      item.type === "TsTypeAliasDeclaration" &&
+      item.id.value === typeName
+    ) {
+      const _tryGetFromInstantiatedParameters = getFromParametersFunc(
+        item.typeParams?.parameters ?? [],
+        ctx,
+      );
+      return {
+        file: path,
+        jsDocSchema: spannableToJsDoc(item),
+        ...await tsTypeToSchemeable(item.typeAnnotation, {
+          ...ctx,
+          tryGetFromInstantiatedParameters: _tryGetFromInstantiatedParameters,
+        }),
+        name: item.id.value,
+      };
+    }
+    if (
+      item.type === "ImportDeclaration"
+    ) {
+      const spec = item.specifiers.find((spec) =>
+        spec.local.value === typeName
+      );
+      if (spec) {
+        try {
+          const pathResolved = import.meta.resolve(
+            path,
+          );
+          const url = pathResolved.startsWith("file")
+            ? fromFileUrl(pathResolved)
+            : pathResolved;
+          const fromTarget = item.source.value.startsWith(".")
+            ? join(
+              dirname(url),
+              item.source.value,
+            )
+            : import.meta.resolve(item.source.value);
+          const newProgram = await parsePath(
+            fromTarget.startsWith("http")
+              ? fromTarget
+              : fromTarget.startsWith("file")
+              ? fromTarget
+              : toFileUrl(fromTarget).toString(),
+          );
+          if (!newProgram) {
+            return UNKNOWN;
+          }
+          return typeNameToSchemeable(
+            (spec as NamedImportSpecifier)?.imported?.value ?? spec.local.value,
+            {
+              ...ctx,
+              path: fromTarget,
+              parsedSource: newProgram,
+            },
+          );
+        } catch (err) {
+          console.log(err, item.source.value, path, import.meta.resolve(path));
+          throw err;
+        }
+      }
+    }
+  }
+  return UNKNOWN;
+};
+
+const wellKnownTypeReferenceToSchemeable = async (
+  ref: TsTypeReference,
+  ctx: SchemeableTransformContext,
 ): Promise<Schemeable | undefined> => {
-  switch (ref.typeName) {
-    case "Partial": {
-      if (ref.typeParams === null || (ref.typeParams?.length ?? 0) < 1) {
+  const typeName = ref.typeName;
+  if (typeName.type !== "Identifier") {
+    return undefined;
+  }
+  const name = typeName.value;
+
+  const typeParams = ref.typeParams?.params ?? [];
+  switch (name) {
+    case "Record": {
+      if (typeParams.length !== 2) {
+        return UNKNOWN;
+      }
+
+      const secondParam = typeParams[1];
+
+      const recordSchemeable = await tsTypeToSchemeable(
+        secondParam,
+        ctx,
+      );
+
+      return {
+        type: "record",
+        name: `record<${recordSchemeable.name}>`,
+        value: recordSchemeable,
+      };
+    }
+    case "PreactComponent": {
+      if (typeParams.length < 1) {
+        return UNKNOWN;
+      }
+      const typeRef = typeParams[0];
+
+      return tsTypeToSchemeable(
+        typeRef,
+        ctx,
+      );
+    }
+    case "Response": {
+      return {
+        type: "inline",
+        name: "Handler",
+        value: {
+          $ref: "#/root/handlers",
+        },
+      };
+    }
+    case "InstanceOf": {
+      if (typeParams.length < 2) {
+        return undefined;
+      }
+      const configName = typeParams[1];
+      if (configName.type !== "TsLiteralType") {
+        return undefined;
+      }
+      const literal = (configName as TsLiteralType).literal;
+      if (literal.type !== "StringLiteral") {
+        return undefined;
+      }
+      return {
+        type: "inline",
+        name: `IF@${btoa(literal.value)}`,
+        value: {
+          $ref: literal.value,
+        },
+      };
+    }
+    case "BlockInstance": {
+      if (typeParams.length < 1) {
+        return undefined;
+      }
+      const configName = typeParams[1];
+      if (configName.type !== "TsLiteralType") {
+        return undefined;
+      }
+      const literal = (configName as TsLiteralType).literal;
+      if (literal.type !== "StringLiteral") {
+        return undefined;
+      }
+      return {
+        type: "inline",
+        name: `BI@${btoa(literal.value)}`,
+        value: {
+          $ref: `#/definitions/${btoa(literal.value)}`,
+        },
+      };
+    }
+    case "Resolvable": {
+      if (typeParams.length < 1) {
         return {
-          type: "unknown",
+          type: "inline",
+          name: "Resolvable",
+          value: {
+            $ref: "#/definitions/Resolvable",
+          },
         };
       }
-      const schemeable = await tsTypeToSchemeableRec(
-        ref.typeParams![0],
-        root,
-        seen,
-        false,
-        typeParams,
-        typeParamsImpl,
+      const typeRef = typeParams[0];
+
+      return tsTypeToSchemeable(
+        typeRef,
+        ctx,
+      );
+    }
+    case "Partial": {
+      if (typeParams.length < 1) {
+        return UNKNOWN;
+      }
+      const schemeable = await tsTypeToSchemeable(
+        typeParams[0],
+        ctx,
       );
       if (schemeable.type !== "object") { // TODO(mcandeia) support arrays, unions and intersections
-        return {
-          type: "unknown",
-        };
+        return UNKNOWN;
       }
 
       const newProperties: ObjectSchemeable["value"] = {};
@@ -117,105 +507,49 @@ const schemeableWellKnownType = async (
       }
       return {
         ...schemeable,
+        name: `Partial@${schemeable.name}`,
         value: newProperties,
-        name: schemeable.name ? `partial${schemeable.name}` : undefined,
-      };
-    }
-    case "Promise": {
-      if (ref.typeParams === null || (ref.typeParams?.length ?? 0) < 1) {
-        return {
-          type: "unknown",
-        };
-      }
-      return tsTypeToSchemeableRec(
-        ref.typeParams![0],
-        root,
-        seen,
-        false,
-        typeParams,
-        typeParamsImpl,
-      );
-    }
-    case "Pick": {
-      if (
-        ref.typeParams === null || (ref.typeParams?.length ?? 0) < 2
-      ) {
-        return {
-          type: "unknown",
-        };
-      }
-      const schemeable = await tsTypeToSchemeableRec(
-        ref.typeParams![0],
-        root,
-        seen,
-        false,
-        typeParams,
-        typeParamsImpl,
-      );
-      const keys = [];
-      if (schemeable.type === "object") { // TODO(mcandeia) support arrays, unions and intersections
-        const newValue: typeof schemeable["value"] = {};
-        const pickKeys = ref.typeParams![1] as TsTypeUnionDef;
-        if (pickKeys?.union) {
-          for (const value of pickKeys?.union) {
-            if (value.kind === "literal" && value.literal.kind === "string") {
-              newValue[value.literal.string] = schemeable
-                .value[value.literal.string];
-              keys.push(value.literal.string);
-            }
-          }
-        }
-        const pickKeysAsLiteral = ref.typeParams![1] as TsTypeDefLiteral;
-        if (
-          pickKeysAsLiteral?.kind &&
-          pickKeysAsLiteral?.literal?.kind === "string"
-        ) {
-          newValue[pickKeysAsLiteral.literal.string] = schemeable
-            .value[pickKeysAsLiteral.literal.string];
-          keys.push(pickKeysAsLiteral.literal.string);
-        }
-        schemeable.value = newValue;
-      }
-      keys.sort();
-      return {
-        ...schemeable,
-        name: `pick${btoa(keys.join())}${schemeable.name}`,
       };
     }
     case "Omit": {
       if (
-        ref.typeParams === null || (ref.typeParams?.length ?? 0) < 2
+        typeParams.length < 2
       ) {
-        return {
-          type: "unknown",
-        };
+        return UNKNOWN;
       }
-      const schemeable = await tsTypeToSchemeableRec(
-        ref.typeParams![0],
-        root,
-        seen,
-        false,
-        typeParams,
-        typeParamsImpl,
+      const schemeable = await tsTypeToSchemeable(
+        typeParams[0],
+        ctx,
       );
       const keys: string[] = [];
       if (schemeable.type === "object") { // TODO(mcandeia) support arrays, unions and intersections
-        const omitKeys = ref.typeParams![1] as TsTypeUnionDef;
-        if (omitKeys?.union) {
-          for (const value of omitKeys?.union) {
-            if (value.kind === "literal" && value.literal.kind === "string") {
-              delete (schemeable.value[value.literal.string]);
-              keys.push(value.literal.string);
+        const omitKeys = typeParams[1] as TsUnionType;
+        if (omitKeys?.types) {
+          for (const value of omitKeys?.types) {
+            if (
+              value.type === "TsLiteralType" &&
+              (value as TsLiteralType).literal.type === "StringLiteral"
+            ) {
+              const val =
+                ((value as TsLiteralType).literal as StringLiteral).value;
+              delete (schemeable
+                .value[
+                  val
+                ]);
+              keys.push(val);
             }
           }
         }
-        const omitKeysAsLiteral = ref.typeParams![1] as TsTypeDefLiteral;
+        const omitKeysAsLiteral = typeParams[1] as TsLiteralType;
         if (
-          omitKeysAsLiteral?.kind &&
-          omitKeysAsLiteral?.literal?.kind === "string"
+          omitKeysAsLiteral?.type === "TsLiteralType" &&
+          (omitKeysAsLiteral?.literal as TsLiteral)?.type === "StringLiteral"
         ) {
-          delete (schemeable.value[omitKeysAsLiteral.literal.string]);
-          keys.push(omitKeysAsLiteral.literal.string);
+          const val =
+            ((omitKeysAsLiteral as TsLiteralType).literal as StringLiteral)
+              .value;
+          delete (schemeable.value[val]);
+          keys.push(val);
         }
       }
       keys.sort();
@@ -226,638 +560,531 @@ const schemeableWellKnownType = async (
           : schemeable.name!,
       };
     }
-    case "LoaderReturnType": {
-      if (ref.typeParams === null || (ref.typeParams?.length ?? 0) < 1) {
-        return {
-          type: "unknown",
-        };
+    case "Pick": {
+      if (
+        typeParams.length < 2
+      ) {
+        return UNKNOWN;
       }
-      return tsTypeToSchemeableRec(
-        ref.typeParams![0],
-        root,
-        seen,
-        false,
-        typeParams,
-        typeParamsImpl,
+      const schemeable = await tsTypeToSchemeable(
+        typeParams[0],
+        ctx,
       );
-    }
-    case "Resolvable": {
-      if (ref.typeParams === null || (ref.typeParams?.length ?? 0) < 1) {
-        return {
-          type: "inline",
-          value: {
-            $ref: "#/definitions/Resolvable",
-          },
-        };
+      const keys: string[] = [];
+      if (schemeable.type === "object") { // TODO(mcandeia) support arrays, unions and intersections
+        const newValue: typeof schemeable["value"] = {};
+        const pickKeys = typeParams[1] as TsUnionType;
+        if (pickKeys?.types) {
+          for (const value of pickKeys?.types) {
+            if (
+              value.type === "TsLiteralType" &&
+              (value as TsLiteralType).literal.type === "StringLiteral"
+            ) {
+              const val =
+                ((value as TsLiteralType).literal as StringLiteral).value;
+              newValue[val] = schemeable.value[val];
+              keys.push(val);
+            }
+          }
+        }
+        const pickKeysAsLiteral = typeParams[1] as TsLiteralType;
+        if (
+          pickKeysAsLiteral?.type === "TsLiteralType" &&
+          (pickKeysAsLiteral?.literal as TsLiteral)?.type === "StringLiteral"
+        ) {
+          const val =
+            ((pickKeysAsLiteral as TsLiteralType).literal as StringLiteral)
+              .value;
+          newValue[val] = schemeable.value[val];
+          keys.push(val);
+        }
       }
-      const typeRef = ref.typeParams![0];
-      if (typeRef.kind !== "typeRef") {
-        return {
-          type: "inline",
-          value: {
-            $ref: "#/definitions/Resolvable",
-          },
-        };
-      }
-
-      return tsTypeToSchemeableRec(
-        typeRef,
-        root,
-        seen,
-        false,
-        typeParams,
-        typeParamsImpl,
-      );
-    }
-    case "BlockInstance": {
-      if ((ref.typeParams?.length ?? 0) < 1) {
-        return undefined;
-      }
-      const configName = ref.typeParams![0];
-      if (configName.kind !== "literal") {
-        return undefined;
-      }
-      const literal = configName.literal;
-      if (literal.kind !== "string") {
-        return undefined;
-      }
+      keys.sort();
       return {
-        type: "inline",
-        value: {
-          $ref: `#/definitions/${btoa(literal.string)}`,
-        },
+        ...schemeable,
+        name: `pick${btoa(keys.join())}${schemeable.name}`,
       };
-    }
-    case "InstanceOf": {
-      if ((ref.typeParams?.length ?? 0) < 2) {
-        return undefined;
-      }
-      const configName = ref.typeParams![1];
-      if (configName.kind !== "literal") {
-        return undefined;
-      }
-      const literal = configName.literal;
-      if (literal.kind !== "string") {
-        return undefined;
-      }
-      return {
-        type: "inline",
-        value: {
-          $ref: literal[literal.kind] as string,
-        },
-      };
-    }
-    case "Response": {
-      return {
-        type: "inline",
-        value: {
-          $ref: "#/root/handlers",
-        },
-      };
-    }
-    case "PreactComponent": {
-      if (ref.typeParams === null || (ref.typeParams?.length ?? 0) < 1) {
-        return {
-          type: "unknown",
-        };
-      }
-      const typeRef = ref.typeParams![0];
-      if (typeRef.kind !== "typeRef") {
-        return {
-          type: "unknown",
-        };
-      }
-
-      return tsTypeToSchemeableRec(
-        typeRef,
-        root,
-        seen,
-        false,
-        typeParams,
-        typeParamsImpl,
-      );
     }
     case "Array": {
-      if (ref.typeParams === null || (ref?.typeParams?.length ?? 0) < 1) {
+      if (typeParams.length < 1) {
         return {
-          name: undefined,
+          name: "unknown[]",
           type: "array",
-          value: {
-            type: "unknown",
-          },
+          value: UNKNOWN,
         };
       }
-      const typeSchemeable = await tsTypeToSchemeableRec(
-        ref.typeParams![0],
-        root,
-        seen,
-        false,
-        typeParams,
-        typeParamsImpl,
+      const typeSchemeable = await tsTypeToSchemeable(
+        typeParams[0],
+        ctx,
       );
 
       return {
-        name: typeSchemeable.name ? `${typeSchemeable.name}[]` : undefined,
         type: "array",
+        name: `${typeSchemeable.name}[]`,
         value: typeSchemeable,
       };
     }
-    case "Record": {
-      if (ref.typeParams?.length !== 2) {
-        return {
-          type: "unknown",
-        };
+    case "Promise": {
+      if (typeParams.length < 1) {
+        return UNKNOWN;
       }
-
-      const recordSchemeable = await tsTypeToSchemeableRec(
-        ref.typeParams[1],
-        root,
-        seen,
-        false,
-        typeParams,
-        typeParamsImpl,
+      return tsTypeToSchemeable(
+        typeParams[0],
+        ctx,
       );
-
-      return {
-        file: recordSchemeable.file,
-        name: recordSchemeable.name
-          ? `${recordSchemeable.name}@record`
-          : undefined,
-        type: "record",
-        value: recordSchemeable,
-      };
     }
-    default:
-      return undefined;
+    case "LoaderReturnType": {
+      if (typeParams.length < 1) {
+        return UNKNOWN;
+      }
+      return tsTypeToSchemeable(
+        typeParams[0],
+        ctx,
+      );
+    }
   }
 };
-
-export const findSchemeableFromNode = async (
-  rootNode: DocNode,
-  root: [string, DocNode[]],
-  seen: Map<DocNode, Schemeable>,
-  typeParams?: Schemeable[],
+// cannot have typeParams but can have type parameters on context
+export const tsTypeToSchemeable = async (
+  tsType: TsType,
+  ctx: SchemeableTransformContext,
+  optional = false,
 ): Promise<Schemeable> => {
-  const currLocation = {
-    file: fromFileUrlOrNoop(rootNode.location.filename),
-    name: rootNode.name,
-  };
-  const ref = {
-    type: "ref",
-    value: currLocation,
-  } as SchemeableRef;
-  const seenValue = seen.get(rootNode);
+  const {
+    path,
+    references,
+  } = ctx;
+  const schemeable = references?.get?.(tsType);
 
-  if (seenValue !== undefined) {
-    return seenValue;
+  if (schemeable !== undefined) {
+    return schemeable;
   }
-  seen.set(rootNode, ref);
-
+  const ref = { type: "unknown", name: "unknown" } as Schemeable;
+  references?.set?.(tsType, ref);
   const resolve = async (): Promise<Schemeable> => {
-    const kind = rootNode.kind;
-    switch (kind) {
-      case "interface": {
-        const allOf = await Promise.all(
-          rootNode.interfaceDef.extends.map(async (tp) => ({
-            ...currLocation,
-            ...await tsTypeToSchemeableRec(tp, root, seen),
-          })),
-        );
+    switch (tsType.type) {
+      case "TsLiteralType": {
+        const type = tsType as TsLiteralType;
+        if (type.literal.type === "TemplateLiteral") {
+          return UNKNOWN;
+        }
+        const literalToJsonSchemaType: Record<string, JSONSchema7TypeName> = {
+          "StringLiteral": "string",
+          "NumericLiteral": "number",
+          "BooleanLiteral": "boolean",
+        };
+        const value = type.literal.value;
+        const constVal = typeof value === "bigint"
+          ? value as unknown as number
+          : value;
         return {
-          ...currLocation,
-          extends: allOf && allOf.length > 0 ? allOf : undefined,
-          type: "object",
-          ...(await typeDefToSchemeable(
-            rootNode.interfaceDef,
-            root,
-            seen,
-            rootNode.interfaceDef.typeParams,
-            typeParams,
-          )),
-          jsDocSchema: rootNode.jsDoc && jsDocToSchema(rootNode.jsDoc),
+          type: "inline",
+          name: `${constVal}`,
+          value: {
+            type: optional
+              ? [literalToJsonSchemaType[type.literal.type], "null"]
+              : literalToJsonSchemaType[type.literal.type], // FIXME(mcandeia) not compliant with JSONSchema
+            const: constVal,
+            default: constVal,
+          },
         };
       }
-      case "typeAlias": {
-        const childSchema = await tsTypeToSchemeableRec(
-          rootNode.typeAliasDef.tsType,
-          root,
-          seen,
-          false,
-          rootNode.typeAliasDef.typeParams,
-          typeParams,
-        );
-
-        return {
-          ...currLocation,
-          ...childSchema,
-          jsDocSchema: rootNode.jsDoc
-            ? { ...childSchema.jsDocSchema, ...jsDocToSchema(rootNode.jsDoc) }
-            : childSchema.jsDocSchema,
-        };
-      }
-      case "import": {
-        const newRoots = await denoDoc(rootNode.importDef.src);
-        const node = newRoots.find((n) => {
-          return n.name === rootNode.importDef.imported;
-        });
-        if (!node) {
+      case "TsIndexedAccessType": {
+        const type = tsType as TsIndexedAccessType;
+        const _indexType = type.indexType;
+        if (_indexType.type !== "TsLiteralType") {
+          return UNKNOWN;
+        }
+        const indexType = (_indexType as TsLiteralType).literal;
+        const schemeable = await tsTypeToSchemeable(type.objectType, ctx);
+        if (
+          schemeable.type === "object" && indexType.type === "StringLiteral"
+        ) {
+          const { [indexType.value]: prop } = schemeable.value;
           return {
-            name: rootNode.name,
-            file: fromFileUrlOrNoop(rootNode.importDef.src),
-            type: "unknown",
+            ...schemeable,
+            name: `${schemeable.name}idx${indexType.value}`,
+            file: path,
+            value: {
+              [indexType.value]: prop,
+            },
           };
         }
+        if (
+          schemeable.type === "array" && indexType.type === "NumericLiteral"
+        ) {
+          const itemSchemeable = Array.isArray(schemeable.value)
+            ? schemeable.value[indexType.value]
+            : schemeable.value;
+          return {
+            ...itemSchemeable,
+            file: path,
+            name: `${schemeable.name}idx${indexType.value}`,
+          };
+        }
+        return UNKNOWN;
+      }
+      case "TsParenthesizedType": {
+        const type = tsType as TsParenthesizedType;
+        return tsTypeToSchemeable(type.typeAnnotation, ctx);
+      }
+      case "TsIntersectionType": {
+        const type = tsType as TsIntersectionType;
+        const value = await Promise.all(
+          type.types.map((tp) => tsTypeToSchemeable(tp, ctx)),
+        );
         return {
-          ...currLocation,
-          ...await findSchemeableFromNode(
-            node,
-            [
-              rootNode.importDef.src,
-              newRoots,
-            ],
-            seen,
-            typeParams,
-          ),
+          file: path,
+          name: value.map((v) => v.name).join("&"),
+          type: "intersection",
+          value,
         };
       }
+      case "TsUnionType": {
+        const type = tsType as TsUnionType;
+        const value = await Promise.all(
+          type.types.map((tp) => tsTypeToSchemeable(tp, ctx)),
+        );
+        return {
+          file: path,
+          type: "union",
+          name: value.map((v) => v.name).join("|"),
+          value,
+        };
+      }
+      case "TsOptionalType": {
+        const type = tsType as TsOptionalType;
+        const genType = await tsTypeToSchemeable(type.typeAnnotation, ctx);
+        return {
+          type: "union",
+          name: `union@null|${genType.name}`,
+          value: [
+            {
+              name: "null",
+              type: "inline",
+              value: {
+                type: "null",
+              },
+            },
+            genType,
+          ],
+        };
+      }
+      case "TsTupleType": {
+        const type = tsType as TsTupleType;
+        const value = await Promise.all(
+          type.elemTypes.map((tp) => tsTypeToSchemeable(tp.ty, ctx)),
+        );
+        return {
+          type: "array",
+          name: `tp@${value.map((v) => v.name).join("|")}`,
+          value,
+        };
+      }
+      case "TsArrayType": {
+        const type = tsType as TsArrayType;
+        const value = await tsTypeToSchemeable(type.elemType, ctx);
+        return {
+          type: "array",
+          name: `${value.name}[]`,
+          value,
+        };
+      }
+      case "TsTypeLiteral": {
+        const type = tsType as TsTypeLiteral;
+        return {
+          file: path,
+          name: `tl@${tsType.span.start}-${tsType.span.end}`,
+          ...await tsTypeElementsToObjectSchemeable(type.members, ctx),
+        };
+      }
+      case "TsKeywordType": {
+        const type = tsType as TsKeywordType;
+        const keywordToType: Record<string, JSONSchema7Type> = {
+          undefined: "null",
+          any: "object",
+          never: "object",
+        };
+
+        if (type.kind === "never") {
+          console.warn(
+            `never keyword is being used on ${path}, falling back to object`,
+          );
+        }
+        const jsonSchemaType = keywordToType[type.kind] ?? type.kind;
+        return {
+          type: "inline",
+          name: `${jsonSchemaType}`,
+          value: type
+            ? ({
+              type: optional ? [jsonSchemaType, "null"] : jsonSchemaType,
+            } as JSONSchema7)
+            : {},
+        };
+      }
+      case "TsTypeReference": {
+        const type = tsType as TsTypeReference;
+        if (type.typeName.type !== "Identifier") {
+          return UNKNOWN;
+        }
+        const wellKnownType = await wellKnownTypeReferenceToSchemeable(
+          type,
+          ctx,
+        );
+
+        if (wellKnownType) {
+          return wellKnownType;
+        }
+        const typeName = type.typeName.value;
+        const parameters = type.typeParams?.params ?? [];
+        const typeParams = await Promise.all(parameters.map((param) => {
+          return tsTypeToSchemeable(param, ctx);
+        }));
+
+        const schemeable = await typeNameToSchemeable(typeName, {
+          ...ctx,
+          instantiatedTypeParams: typeParams,
+        });
+
+        return {
+          ...schemeable,
+          name: `${schemeable.name}${
+            typeParams.length > 0
+              ? `+${typeParams.map((p) => p.name).join("+")}`
+              : ""
+          }`,
+        };
+      }
+      default:
+        return UNKNOWN;
     }
-    return {
-      ...currLocation,
-      type: "unknown",
-    };
   };
-  const resolved = await resolve();
-  ref.value = resolved;
-  seen.delete(rootNode);
+
+  const resolved: Schemeable = await resolve();
+  Object.assign(ref, resolved);
+  references?.delete?.(tsType);
   return resolved;
 };
 
-const typeDefToSchemeable = async (
-  node: InterfaceDef | TsTypeLiteralDef,
-  root: [string, DocNode[]],
-  seen: Map<DocNode, Schemeable>,
-  typeParams?: TsTypeParamDef[],
-  typeParamsImpl?: Schemeable[],
-): Promise<Omit<ObjectSchemeable, "id" | "type">> => {
-  const properties = await Promise.all(
-    node.properties.map(async (property) => {
-      const jsDocSchema =
-        (property as InterfaceDef["properties"][number]).jsDoc &&
-        jsDocToSchema((property as InterfaceDef["properties"][number]).jsDoc!);
-      const schema = await tsTypeToSchemeableRec(
-        property.tsType!,
-        root,
-        new Map(seen),
-        property.optional,
-        typeParams,
-        typeParamsImpl,
-      );
-
-      return [
-        property.name,
-        {
-          required: !property.optional,
-          schemeable: schema,
-          jsDocSchema: jsDocSchema,
-          title: beautify(property.name),
-        },
-      ];
-    }),
-  );
-
-  return {
-    value: Object.fromEntries(properties),
-  };
-};
-export const tsTypeToSchemeable = async (
-  rootNode: DocNode,
-  node: TsTypeDef,
-  root: [string, DocNode[]],
-  optional?: boolean,
-): Promise<Schemeable> => {
-  const seen = new Map();
-  seen.set(rootNode, true);
-  const resp = await tsTypeToSchemeableRec(node, root, seen, optional);
-  seen.delete(rootNode);
-  return resp;
-};
-
-const tsTypeToSchemeableRec = async (
-  node: TsTypeDef,
-  root: [string, DocNode[]],
-  seen: Map<DocNode, Schemeable>,
-  optional?: boolean,
-  typeParams?: TsTypeParamDef[],
-  typeParamsImpl?: Schemeable[],
-): Promise<Schemeable> => {
-  const kind = node.kind;
-  switch (kind) {
-    case "parenthesized": {
-      return await tsTypeToSchemeableRec(
-        node.parenthesized,
-        root,
-        seen,
-        optional,
-        typeParams,
-        typeParamsImpl,
-      );
+export interface FunctionCanonicalDeclaration {
+  exp: FunctionExpression | FunctionDeclaration | ArrowFunctionExpression;
+  path: string;
+  parsedSource: ParsedSource;
+  jsDoc: JSONSchema7;
+}
+const findFuncFromExportNamedDeclaration = async (
+  funcName: string,
+  item: ExportNamedDeclaration,
+  path: string,
+): Promise<FunctionCanonicalDeclaration | undefined> => {
+  for (const spec of item.specifiers) {
+    if (
+      item.source !== undefined &&
+      spec.type === "ExportSpecifier" &&
+      (spec.exported?.value ?? spec.orig.value) === funcName
+    ) {
+      const url = item.source.value.startsWith(".")
+        ? import.meta.resolve(
+          new URL(item.source.value, import.meta.resolve(path)).toString(),
+        )
+        : import.meta.resolve(item.source.value);
+      const isFromDefault = spec.orig.value === "default";
+      const newProgram = await parsePath(url);
+      if (!newProgram) {
+        return undefined;
+      }
+      if (isFromDefault) {
+        return findDefaultFuncExport(url, newProgram);
+      } else {
+        return findFuncExport(spec.orig.value, url, newProgram);
+      }
     }
-    case "indexedAccess": {
-      const objSchemeable = await tsTypeToSchemeableRec(
-        node.indexedAccess.objType,
-        root,
-        seen,
-        optional,
-        typeParams,
-        typeParamsImpl,
-      );
-      const indexType = node.indexedAccess.indexType;
-      if (indexType.kind !== "literal") { // supporting only literal access A["b"]
-        return {
-          type: "unknown",
-        };
-      }
-      if (
-        objSchemeable.type === "object" && indexType.literal.kind === "string"
-      ) {
-        const { [indexType.literal.string]: prop } = objSchemeable.value;
-        return {
-          ...objSchemeable,
-          name: objSchemeable.name
-            ? `indexedAccess${indexType.literal.string}${objSchemeable.name}`
-            : undefined,
-          value: {
-            [indexType.literal.string]: prop,
-          },
-        };
-      }
-
-      if (
-        objSchemeable.type === "array" && indexType.literal.kind === "number" &&
-        !Array.isArray(objSchemeable.value)
-      ) {
-        const type = objSchemeable.value;
-        return {
-          ...type,
-          name: type.name
-            ? `indexedAccess${indexType.literal.number}${type.name}`
-            : undefined,
-        };
-      }
-      return {
-        type: "unknown",
-      };
-    }
-    case "array": {
-      const typeSchemeable = await tsTypeToSchemeableRec(
-        node.array,
-        root,
-        seen,
-        optional,
-        typeParams,
-        typeParamsImpl,
-      );
-
-      return {
-        name: typeSchemeable.name ? `${typeSchemeable.name}[]` : undefined,
-        file: typeSchemeable.file,
-        type: "array",
-        value: typeSchemeable,
-      };
-    }
-    case "typeRef": {
-      const wellknown = await schemeableWellKnownType(
-        node.typeRef,
-        root,
-        seen,
-        typeParams,
-        typeParamsImpl,
-      );
-      if (wellknown) {
-        return wellknown;
-      }
-      const rootNode = root[1].find((n) => {
-        return n.name === node.typeRef.typeName;
-      });
-      if (!rootNode) {
-        const typeParamsMap = (typeParams ?? []).reduce(
-          (
-            tpParamsMap,
-            tpParam,
-            i,
-          ) => {
-            return {
-              ...tpParamsMap,
-              [tpParam.name]: [tpParam.default, i] as [
-                TsTypeDef | undefined,
-                number,
-              ],
-            };
-          },
-          {} as Record<string, [TsTypeDef | undefined, number]>,
-        );
-        if (node.typeRef.typeName in typeParamsMap) {
-          const anchor = node.typeRef.typeName;
-          const [genType, idx] = typeParamsMap[node.typeRef.typeName];
-          const fromTypeParams = (typeParamsImpl ?? [])[idx];
-          if (fromTypeParams) {
-            return {
-              ...fromTypeParams,
-              anchor,
-            };
-          }
-          if (genType) {
-            const typeSchemeable = await tsTypeToSchemeableRec(
-              genType,
-              root,
-              seen,
-              optional,
-              typeParams,
-              typeParamsImpl,
-            );
-            return {
-              ...typeSchemeable,
-              anchor,
-            };
-          }
-          const anyDoc = root[1]?.[0];
-          return {
-            type: "inline",
-            anchor,
-            file: anyDoc
-              ? fromFileUrlOrNoop(anyDoc?.location.filename)
-              : undefined,
-            value: {} as JSONSchema7,
-          };
-        }
-        return {
-          type: "unknown",
-        };
-      }
-
-      const implementations = await Promise.all(
-        (node.typeRef.typeParams ?? []).map((impl) =>
-          tsTypeToSchemeableRec(
-            impl,
-            root,
-            seen,
-            optional,
-            typeParams,
-            typeParamsImpl,
-          )
-        ),
-      );
-      const withName = (implementations ?? []).map((impl) => impl.name).filter(
-        Boolean,
-      ).join("+");
-      const resp = await findSchemeableFromNode(
-        rootNode,
-        root,
-        seen,
-        implementations,
-      );
-      return {
-        ...resp,
-        name: resp.name && withName.length > 0
-          ? `${resp.name}+${withName}`
-          : resp.name,
-      };
-    }
-    case "keyword": {
-      if (node.keyword === "unknown") {
-        return {
-          type: "unknown",
-        };
-      }
-      const keywordToType: Record<string, JSONSchema7Type> = {
-        undefined: "null",
-        any: "object",
-        never: "object",
-      };
-      if (node.keyword === "never") {
-        console.warn(
-          `never keyword is being used on ${root[0]}, falling back to object`,
-        );
-      }
-      const type = keywordToType[node.keyword] ?? node.keyword;
-      return {
-        type: "inline",
-        value: type
-          ? ({
-            type: optional ? [type, "null"] : type,
-          } as JSONSchema7)
-          : {},
-      };
-    }
-    case "typeLiteral":
-      return {
-        type: "object",
-        ...(await typeDefToSchemeable(
-          node.typeLiteral,
-          root,
-          seen,
-          typeParams,
-          typeParamsImpl,
-        )),
-      };
-    case "literal": {
-      // deno-lint-ignore no-explicit-any
-      const value = (node.literal as any)[node.literal.kind];
-      return {
-        type: "inline",
-        value: {
-          type: node.literal.kind as JSONSchema7TypeName, // FIXME(mcandeia) not compliant with JSONSchema
-          const: value,
-          default: value,
-        },
-      };
-    }
-    case "intersection": {
-      const values = await Promise.all(
-        node.intersection.map((t) =>
-          tsTypeToSchemeableRec(
-            t,
-            root,
-            new Map(seen),
-            optional,
-            typeParams,
-            typeParamsImpl,
-          )
-        ),
-      );
-      const ids = [];
-      for (let i = 0; i < node.intersection.length; i++) {
-        const tp = values[i];
-        if (tp?.name) {
-          ids.push(tp.name);
-        } else if (tp && tp.type === "inline" && tp.value.type === "null") {
-          ids.push("null");
-        } else if (node.repr) {
-          ids.push(node.repr);
-        } else if (tp.anchor) {
-          ids.push(tp.anchor);
-        } else {
-          const hash = await crypto.subtle.digest(
-            "MD5",
-            new TextEncoder().encode(JSON.stringify(tp)),
-          );
-          ids.push(toHashString(hash));
-        }
-      }
-      ids.sort();
-      const intersectionTypeId = ids.length === 0 ? undefined : ids.join("&");
-      return {
-        name: intersectionTypeId,
-        file: values.find((v) => v.file)?.file,
-        value: values,
-        type: "intersection",
-      };
-    }
-    case "union": {
-      const values = await Promise.all(
-        node.union.map((t) =>
-          tsTypeToSchemeableRec(
-            t,
-            root,
-            new Map(seen),
-            false,
-            typeParams,
-            typeParamsImpl,
-          )
-        ),
-      );
-      const ids = [];
-      for (let i = 0; i < node.union.length; i++) {
-        const tp = values[i];
-        if (tp?.name) {
-          ids.push(tp.name);
-        } else if (tp && tp.type === "inline" && tp.value.type === "null") {
-          ids.push("null");
-        } else if (node.repr) {
-          ids.push(node.repr);
-        } else if (tp.anchor) {
-          ids.push(tp.anchor);
-        } else {
-          const hash = await crypto.subtle.digest(
-            "MD5",
-            new TextEncoder().encode(JSON.stringify(tp)),
-          );
-          ids.push(toHashString(hash));
-        }
-      }
-      ids.sort();
-      const unionTypeId = ids.length === 0 ? undefined : ids.join("|");
-      return {
-        name: unionTypeId,
-        file: values.find((v) => v.file)?.file,
-        value: values,
-        type: "union",
-      };
-    }
-    case "fnOrConstructor":
-    default:
-      return {
-        type: "unknown",
-      };
   }
+  return undefined;
+};
+
+const findFunc = async (
+  funcName: string,
+  path: string,
+  parsedSource: ParsedSource,
+): Promise<[FunctionCanonicalDeclaration, boolean] | undefined> => {
+  for (const item of parsedSource.program.body) {
+    if (item.type === "ExportNamedDeclaration") {
+      const found = await findFuncFromExportNamedDeclaration(
+        funcName,
+        item,
+        path,
+      );
+      if (found) {
+        return [found, true];
+      }
+    }
+
+    if (
+      item.type === "FunctionDeclaration" &&
+      item.identifier.value === funcName
+    ) {
+      return [{
+        path,
+        parsedSource,
+        exp: item,
+        jsDoc: spannableToJsDoc(item),
+      }, false];
+    }
+
+    if (
+      item.type === "VariableDeclaration"
+    ) {
+      for (const decl of item.declarations) {
+        if (
+          decl.id.type === "Identifier" && decl.id.value === funcName &&
+          decl.init && decl.init.type === "ArrowFunctionExpression"
+        ) {
+          return [{
+            path,
+            parsedSource,
+            exp: decl.init,
+            jsDoc: spannableToJsDoc(item),
+          }, false];
+        }
+      }
+    }
+
+    if (
+      item.type === "ExportDeclaration" &&
+      item.declaration.type === "FunctionDeclaration" &&
+      item.declaration.identifier.value === funcName
+    ) {
+      return [{
+        path,
+        parsedSource,
+        exp: item.declaration,
+        jsDoc: spannableToJsDoc(item),
+      }, true];
+    }
+
+    if (
+      item.type === "ExportDeclaration" &&
+      item.declaration.type === "VariableDeclaration"
+    ) {
+      for (const decl of item.declaration.declarations) {
+        if (
+          decl.id.type === "Identifier" && decl.id.value === funcName &&
+          decl.init && decl.init.type === "ArrowFunctionExpression"
+        ) {
+          return [{
+            path,
+            parsedSource,
+            exp: decl.init,
+            jsDoc: spannableToJsDoc(item),
+          }, true];
+        }
+      }
+    }
+  }
+};
+const findFuncExport = async (
+  funcName: string,
+  path: string,
+  program: ParsedSource,
+): Promise<FunctionCanonicalDeclaration | undefined> => {
+  const func = await findFunc(funcName, path, program);
+  if (!func) {
+    return undefined;
+  }
+  return func[1] ? func[0] : undefined;
+};
+
+export const findDefaultFuncExport = async (
+  path: string,
+  parsedSource: ParsedSource,
+): Promise<FunctionCanonicalDeclaration | undefined> => {
+  for (const item of parsedSource.program.body) {
+    if (
+      item.type === "ExportDefaultExpression" &&
+      item.expression.type === "Identifier"
+    ) {
+      const func = await findFunc(item.expression.value, path, parsedSource);
+      return func?.[0];
+    }
+    if (
+      item.type === "ExportDefaultDeclaration" &&
+      item.decl.type === "FunctionExpression"
+    ) {
+      return {
+        exp: item.decl,
+        path,
+        parsedSource,
+        jsDoc: spannableToJsDoc(item),
+      };
+    }
+    if (item.type === "ExportNamedDeclaration") {
+      const found = await findFuncFromExportNamedDeclaration(
+        "default",
+        item,
+        path,
+      );
+      if (found) {
+        return found;
+      }
+    }
+  }
+  return undefined;
+};
+export const programToBlockRef = async (
+  _path: string,
+  _program: ParsedSource,
+  schemeableReferences?: Map<TsType, Schemeable>,
+  introspect?: IntrospectParams,
+): Promise<BlockModuleRef | undefined> => {
+  const funcNames = introspect?.funcNames ?? ["default"];
+  for (const name of funcNames) {
+    const fn = name === "default"
+      ? await findDefaultFuncExport(_path, _program)
+      : await findFuncExport(name, _path, _program);
+    if (!fn) {
+      continue;
+    }
+
+    const includeReturn = introspect?.includeReturn;
+    const { exp: func, path, parsedSource } = fn;
+
+    const retn = typeof includeReturn === "function"
+      ? func.returnType
+        ? includeReturn(func.returnType.typeAnnotation)
+        : undefined
+      : func.returnType?.typeAnnotation;
+
+    const baseBlockRef = {
+      functionJSDoc: fn.jsDoc, //func.jsDoc && jsDocToSchema(func.jsDoc),
+      functionRef: path,
+      outputSchema: includeReturn && retn
+        ? await tsTypeToSchemeable(retn, {
+          path,
+          parsedSource,
+          references: schemeableReferences,
+        })
+        : undefined,
+    };
+    const paramIdx = 0;
+    if (func.params.length === 0) {
+      return baseBlockRef;
+    }
+    const param = func.params[paramIdx];
+    const pat = (param as Param)?.pat ?? param as Pattern;
+    const typeAnnotation = (pat as { typeAnnotation?: TsTypeAnnotation })
+      ?.typeAnnotation?.typeAnnotation;
+    if (!typeAnnotation) {
+      return baseBlockRef;
+    }
+    return {
+      ...baseBlockRef,
+      inputSchema: await tsTypeToSchemeable(typeAnnotation, {
+        path,
+        parsedSource,
+        references: schemeableReferences,
+      }),
+    };
+  }
+  return undefined;
 };
