+++ mod.ts
@@ -13,4 +13,3 @@ export type { WorkflowGen } from "./deps.ts";
 export type { App, AppContext, AppModule, AppRuntime } from "./types.ts";
 export type { StreamProps } from "./utils/invoke.ts";
 export { stylesPlugin };
-
